renames source geodatabase resolves #3

+++ src/py/publisher.py
@@ -1,5 +1,6 @@
 import organization
 import os
+import stat
 import shutil
 
 # we may manage several types of content
@@ -40,14 +41,19 @@ class localgdb(object):
         if os.path.isdir(filegdb):
             self.gdb  = filegdb
             self.gdbname = os.path.basename(self.gdb)
+            self.gdbpath = os.path.dirname(self.gdb)
         else:
             raise ValueError("{0} is not a valid path to a file geodatabase".format(filegdb))
-                
+
+        self.renamed = None
+        self.zipped  = None
+        
     def zip(self
            ,zippath):
         
         # plan to always zip to a new path
         # the source cscl environment is shared
+        # zip then move for performance
         
         # zippath C:\temp\dir2
         # C:\temp\dir1\mydata.gdb
@@ -62,9 +68,50 @@ class localgdb(object):
 
         shutil.move("{0}.zip".format(self.gdb)
                    ,self.zipped)
+
+    def renamezip(self
+                 ,zippath
+                 ,name):
+
+        # 1 copy 2 rename 3 zip
+        self.renamed = os.path.join(zippath,"{0}".format(name))
+        self.zipped = os.path.join(zippath
+                                  ,"{0}.zip".format(name))
+
+        # C:\temp\dir1\mydata.gdb to
+        # C:\temp\dir2\mydata.gdb
+        shutil.copytree(self.gdb
+                       ,os.path.join(zippath,"{0}".format(self.gdbname)))
+
+        # C:\temp\dir2\mydata.gdb to
+        # C:\temp\dir2\pubdata.gdb
+        os.rename(os.path.join(zippath,"{0}".format(self.gdbname))
+                 ,self.renamed)
         
+        # C:\temp\dir2\pubdata.gdb
+        # C:\temp\dir2\pubdata.gdb.zip
+        shutil.make_archive(self.renamed
+                           ,'zip'
+                           ,self.renamed)
+
+    def remove_readonly(self
+                       ,func
+                       ,path
+                       ,_):
+        
+        os.chmod(path, stat.S_IWRITE)
+        func(path)
+
     def clean(self):
         
-        os.remove(self.zipped)
+        if (self.renamed and os.path.exists(self.renamed)):
+            # PermissionError: [WinError 5] 
+            #     Access is denied: 'D:\\temp\\renamesample.gdb'
+            # some files in the gdb are readonly (but not the dir itself)
+            # onexc callback is copied verbatim from the shutil docs
+            shutil.rmtree(self.renamed, onerror=self.remove_readonly)
+
+        if self.zipped:
+            os.remove(self.zipped)
